% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{as.character.ustring}
\alias{as.character.ustring}
\title{Convert a ustring to text}
\usage{
\method{as.character}{ustring}(x, ...)
}
\arguments{
\item{x}{a ustring}

\item{...}{ignored}
}
\description{
Convert a ustring to text
}
