% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{as.character.utfstring}
\alias{as.character.utfstring}
\title{Convert a utfstring to text}
\usage{
\method{as.character}{utfstring}(x, ...)
}
\arguments{
\item{x}{a utfstring}

\item{...}{ignored}
}
\description{
Convert a utfstring to text
}
