% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{as.ustring.character}
\alias{as.ustring.character}
\alias{as.ustring.ustring}
\alias{as.ustring}
\title{Convert a scaler character to ustring}
\usage{
\method{as.ustring}{character}(x, encoding, ...)

\method{as.ustring}{ustring}(x, encoding, ...)

as.ustring(x, ...)
}
\arguments{
\item{x}{a scalar character}

\item{encoding}{a scalar character}

\item{...}{ignored}
}
\description{
Convert a scaler character to ustring

Convert a scaler character to ustring
}
