% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{as.utfstring.character}
\alias{as.utfstring.character}
\title{Convert a scaler character to utfstring}
\usage{
\method{as.utfstring}{character}(x, encoding)
}
\arguments{
\item{x}{a scalar character}

\item{encoding}{a scalar character}
}
\description{
Convert a scaler character to utfstring
}
