% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{code_point}
\alias{code_point}
\title{Get the unicode code for each code point in a UTF-8 encoded text.}
\usage{
code_point(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Get the unicode code for each code point in a UTF-8 encoded text.
}
