% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{code_points_to_text}
\alias{code_points_to_text}
\title{Convert unicode code points to text}
\usage{
code_points_to_text(x)
}
\arguments{
\item{x}{a vector of code points}
}
\value{
a scalar character
}
\description{
Convert unicode code points to text
}
