% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{code_points_to_utf32}
\alias{code_points_to_utf32}
\title{Convert unicode code points to UTF-32 ustring.}
\usage{
code_points_to_utf32(x, endian = "big")
}
\arguments{
\item{x}{a vector of code points}

\item{endian}{little endian or big endian?}
}
\value{
a scalar character
}
\description{
Convert unicode code points to UTF-32 ustring.
}
