% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{code_points_to_utf8}
\alias{code_points_to_utf8}
\title{Convert unicode code points to UTF-8 ustring.}
\usage{
code_points_to_utf8(x)
}
\arguments{
\item{x}{a vector of code points}
}
\value{
a scalar character
}
\description{
Convert unicode code points to UTF-8 ustring.
}
