% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{n_code_points}
\alias{n_code_points}
\title{Calculate the number of code points. Fasterer version of \code{nchar(..., type = "chars")}.}
\usage{
n_code_points(x)
}
\arguments{
\item{x}{a scalar character or ustring}
}
\description{
Calculate the number of code points. Fasterer version of \code{nchar(..., type = "chars")}.
}
