% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{text_to_utf16}
\alias{text_to_utf16}
\title{Convert a UTF-8 text to UTF-16 utfstring.}
\usage{
text_to_utf16(text, endian = "little")
}
\arguments{
\item{text}{a scalar character}

\item{endian}{little endian or big endian?}
}
\description{
Convert a UTF-8 text to UTF-16 utfstring.
}
