% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{text_to_utf32}
\alias{text_to_utf32}
\title{Convert UTF-8 text to UTF-32 ustring.}
\usage{
text_to_utf32(text, endian = "big")
}
\arguments{
\item{text}{UTF-8 ustring}

\item{endian}{little endian or big endian?}
}
\description{
Convert UTF-8 text to UTF-32 ustring.
}
