% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{text_to_utf8}
\alias{text_to_utf8}
\title{Convert a UTF-8 text to UTF-8 utfstring.}
\usage{
text_to_utf8(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Convert a UTF-8 text to UTF-8 utfstring.
}
