% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\docType{package}
\name{ustring-package}
\alias{ustring}
\alias{ustring-package}
\title{ustring: Tools For Unicode Strings}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
