% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf16_len}
\alias{utf16_len}
\title{Calcuate the number of code points of a UTF-16 raw string.}
\usage{
utf16_len(text)
}
\arguments{
\item{text}{a UTF-16 raw string}
}
\description{
Calcuate the number of code points of a UTF-16 raw string.
}
