% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{utf16_to_text}
\alias{utf16_to_text}
\title{Convert a UTF-16 ustring to UTF-8 text.}
\usage{
utf16_to_text(s)
}
\arguments{
\item{s}{a UTF-16 ustring}
}
\description{
Convert a UTF-16 ustring to UTF-8 text.
}
