% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf32_len}
\alias{utf32_len}
\title{Calcuate the number of code points of a UTF-32 raw string.}
\usage{
utf32_len(text)
}
\arguments{
\item{text}{a UTF-32 raw string}
}
\description{
Calcuate the number of code points of a UTF-32 raw string.
}
