% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf32_length}
\alias{utf32_length}
\title{Calcuate the number of code points of a UTF-32 raw string.}
\usage{
utf32_length(text)
}
\arguments{
\item{text}{a UTF-32 raw string}
}
\description{
Calcuate the number of code points of a UTF-32 raw string.
}
