% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustring.R
\name{utf32_to_text}
\alias{utf32_to_text}
\title{Convert a UTF-32 ustring to UTF-8 text.}
\usage{
utf32_to_text(s)
}
\arguments{
\item{s}{a UTF-32 ustring}
}
\description{
Convert a UTF-32 ustring to UTF-8 text.
}
