% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf32_to_utf8}
\alias{utf32_to_utf8}
\title{Convert a UTF-32 raw string to UTF-8 string.}
\usage{
utf32_to_utf8(s)
}
\arguments{
\item{s}{a UTF-32 raw string}
}
\description{
Convert a UTF-32 raw string to UTF-8 string.
}
