% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_code}
\alias{utf8_code}
\title{Get the unicode code for each code point in a UTF-8 encoded text.}
\usage{
utf8_code(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Get the unicode code for each code point in a UTF-8 encoded text.
}
