% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_code_nbytes}
\alias{utf8_code_nbytes}
\title{Calcuate number of bytes of each code point in a UTF-8 encoded text.}
\usage{
utf8_code_nbytes(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Calcuate number of bytes of each code point in a UTF-8 encoded text.
}
