% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_codelen}
\alias{utf8_codelen}
\title{Calcuate code point lenghts of each char in a UTF-8 encoded text.}
\usage{
utf8_codelen(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Calcuate code point lenghts of each char in a UTF-8 encoded text.
}
