% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_decode}
\alias{utf8_decode}
\title{Get the unicode code point for each char in a UTF-8 encoded text.}
\usage{
utf8_decode(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Get the unicode code point for each char in a UTF-8 encoded text.
}
