% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_length}
\alias{utf8_length}
\title{Calcuate the number of code points of a UTF-8 encoded text.}
\usage{
utf8_length(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Calcuate the number of code points of a UTF-8 encoded text.
}
