% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_to_text}
\alias{utf8_to_text}
\title{Convert a UTF-8 text to UTF-8 utfstring.}
\usage{
utf8_to_text(s)
}
\arguments{
\item{s}{a UTF-8 utfstring}
}
\description{
Convert a UTF-8 text to UTF-8 utfstring.
}
