% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_to_utf16}
\alias{utf8_to_utf16}
\title{Convert a UTF-8 string to UTF-16 raw string.}
\usage{
utf8_to_utf16(text, endian = c("little", "big"))
}
\arguments{
\item{text}{a scalar character}

\item{endian}{little endian or big endian?}
}
\description{
Convert a UTF-8 string to UTF-16 raw string.
}
