% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_to_utf32}
\alias{utf8_to_utf32}
\title{Convert a UTF-8 encoded text to UTF-32 bytestring.}
\usage{
utf8_to_utf32(text)
}
\arguments{
\item{text}{a scalar character}
}
\description{
Convert a UTF-8 encoded text to UTF-32 bytestring.
}
