% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\docType{package}
\name{utftools-package}
\alias{utftools}
\alias{utftools-package}
\title{utftools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
