% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{text_to_utf32}
\alias{text_to_utf32}
\title{Convert UTF-8 text to UTF-32 utfstring.}
\usage{
text_to_utf32(text, endian = "little")
}
\arguments{
\item{text}{UTF-8 utfstring}

\item{endian}{little endian or big endian?}
}
\description{
Convert UTF-8 text to UTF-32 utfstring.
}
