% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf16_to_utf8}
\alias{utf16_to_utf8}
\title{Convert a UTF-16 utfstring to UTF-8 utfstring.}
\usage{
utf16_to_utf8(s)
}
\arguments{
\item{s}{a UTF-16 utfstring}
}
\description{
Convert a UTF-16 utfstring to UTF-8 utfstring.
}
