% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\name{utf8_to_utf32}
\alias{utf8_to_utf32}
\title{Convert a UTF-8 string to UTF-32 raw string.}
\usage{
utf8_to_utf32(text, endian = "little")
}
\arguments{
\item{text}{a scalar character}

\item{endian}{little endian or big endian?}
}
\description{
Convert a UTF-8 string to UTF-32 raw string.
}
