% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utftools.R
\docType{package}
\name{utftools-package}
\alias{utftools}
\alias{utftools-package}
\title{utftools: Some Unicode Tools}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
